log all in deferred log for resolver when level is verbose
When the logging level is Verbose, we should enable verbose logging in the
deferred log for the resolver in resolve() for build plugins.

This enables debugging when PluginBuild Resolve fails when called within a build
plugin using the Go api.

Set the log level of the deferred logger to match the parent logger.

Signed-off-by: Christian Stewart <christian@aperture.us>