fix(logger): defer log constructor should accept log level
The NewDeferLog constructor defaults to LogLevelInfo regardless of the log level
specified in the bundler options. We should instead inherit the log level from
the parent logger so that we don't drop debug messages.

Add a new logLevel parameter to NewDeferLog. Updated all locations where
NewDeferLog is called to either pass the parent logger level or to pass the old
default value of LogLevelInfo if there is no parent logger.

Signed-off-by: Christian Stewart <christian@aperture.us>